refactor :  api 공용 모듈을 분리
