+++ src/lib/api/core.ts
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export const api = axios.create({
+  baseURL: import.meta.env.VITE_API_BASE_URL,
+});
