+++ src/lib/api/index.ts
@@ -1,0 +1,2 @@
+export * from './rankingProducts';
+export * from './themes';
